+++ lib/agent0/agent0/base/agent/eth_agent.py
@@ -10,9 +10,9 @@ from fixedpointmath import FixedPoint
 from hexbytes import HexBytes
 from web3 import Web3
 
-from agent0.base import Quantity, TokenType, Trade
 from agent0.base.policies import BasePolicy, NoActionPolicy
 
+from ..types import Quantity, TokenType, Trade
 from .eth_wallet import EthWallet
 
 Policy = TypeVar("Policy", bound=BasePolicy)
