+++ lib/agent0/agent0/hyperdrive/policies/random_test.py
@@ -25,8 +25,7 @@ from agent0.hyperdrive.policies import PolicyZoo
 
 if TYPE_CHECKING:
     from ethpy.hyperdrive import HyperdriveAddresses
-
-    from agent0.hyperdrive.test_fixtures.local_chain import DeployedHyperdrivePool
+    from ethpy.test_fixtures import DeployedHyperdrivePool
 
 # pylint: disable=too-many-locals
