--- lib/agent0/agent0/hyperdrive/test_fixtures/__init__.py
@@ -1,9 +1,0 @@
-"""Test fixtures for ethpy"""
-
-from .local_chain import (
-    hyperdrive_read_interface,
-    hyperdrive_read_write_interface,
-    init_local_hyperdrive_pool,
-    local_chain,
-    local_hyperdrive_pool,
-)
