+++ lib/agent0/agent0/test_fixtures/__init__.py
@@ -1,3 +1,4 @@
 """Test fixtures for agent0."""
 
 from .chain import chain
+from .interface import hyperdrive_read_interface, hyperdrive_read_write_interface
