+++ lib/ethpy/ethpy/test_fixtures/deployed_pool.py
@@ -2,65 +2,18 @@
 
 from __future__ import annotations
 
-import subprocess
-import time
-from typing import Callable, Iterator, cast
+from typing import Callable, Iterator
 
 import pytest
-from eth_typing import URI
 from ethpy.base import initialize_web3_with_http_provider
-from ethpy.eth_config import EthConfig
-from ethpy.hyperdrive import DeployedHyperdrivePool, HyperdriveAddresses, deploy_hyperdrive_from_factory
+from ethpy.hyperdrive import DeployedHyperdrivePool, deploy_hyperdrive_from_factory
 from fixedpointmath import FixedPoint
 from hypertypes import FactoryConfig, Fees, PoolDeployConfig
-from web3 import HTTPProvider
 from web3.constants import ADDRESS_ZERO
 from web3.types import RPCEndpoint
 
-from agent0.hyperdrive import HyperdriveReadInterface, HyperdriveReadWriteInterface
-
 # we need to use the outer name for fixtures
 # pylint: disable=redefined-outer-name
-
-
-def launch_local_chain(anvil_port: int = 9999, host: str = "127.0.0.1") -> Iterator[str]:
-    """Launch a local anvil chain.
-
-    Arguments
-    ---------
-    anvil_port: int
-        Port number for the anvil chain.
-    host: str
-        Host address.
-
-    Yields
-    ------
-    str
-        The local anvil chain URI.
-    """
-    # Supress output of anvil
-    anvil_process = subprocess.Popen(  # pylint: disable=consider-using-with
-        ["anvil", "--host", host, "--port", str(anvil_port), "--code-size-limit", "9999999999"],
-        stdout=subprocess.DEVNULL,
-        stderr=subprocess.STDOUT,
-    )
-    time.sleep(3)  # Wait for anvil chain to initialize
-
-    yield f"http://{host}:{anvil_port}"
-    anvil_process.kill()  # Kill anvil process at end
-
-
-@pytest.fixture(scope="session")
-def local_chain() -> Iterator[str]:
-    """Fixture representing a local anvil chain.
-    This fixture is session scoped, so each test will deploy its own pool for testing.
-
-    Yields
-    ------
-    str
-        The local anvil chain URI.
-    """
-    yield from launch_local_chain()
 
 
 @pytest.fixture(scope="session")
@@ -266,73 +219,3 @@ def local_hyperdrive_pool(
 
     # Revert to snapshot
     reset(snapshot_id)
-
-
-def create_hyperdrive_read_interface(_local_hyperdrive_pool: DeployedHyperdrivePool) -> HyperdriveReadInterface:
-    """Set up the hyperdrive read interface to access a deployed hyperdrive pool.
-
-    All arguments are fixtures.
-
-    Returns
-    -------
-        HyperdriveReadInterface
-            The interface to access the deployed hyperdrive pool.
-    """
-    rpc_uri = cast(HTTPProvider, _local_hyperdrive_pool.web3.provider).endpoint_uri or URI("http://localhost:8545")
-    hyperdrive_contract_addresses: HyperdriveAddresses = _local_hyperdrive_pool.hyperdrive_contract_addresses
-    eth_config = EthConfig(artifacts_uri="not used", rpc_uri=rpc_uri, abi_dir="./packages/hyperdrive/src/abis")
-    return HyperdriveReadInterface(eth_config, addresses=hyperdrive_contract_addresses)
-
-
-@pytest.fixture(scope="function")
-def hyperdrive_read_interface(local_hyperdrive_pool: DeployedHyperdrivePool) -> Iterator[HyperdriveReadInterface]:
-    """Fixture representing a hyperdrive read interface to a deployed hyperdrive pool.
-
-    Arguments
-    ---------
-    local_hyperdrive_pool: DeployedHyperdrivePool
-        Fixture representing the deployed hyperdrive pool
-
-    Yields
-    ------
-    HyperdriveReadInterface
-        The interface to access the deployed hyperdrive pool.
-    """
-    yield create_hyperdrive_read_interface(local_hyperdrive_pool)
-
-
-def create_hyperdrive_read_write_interface(
-    _local_hyperdrive_pool: DeployedHyperdrivePool,
-) -> HyperdriveReadWriteInterface:
-    """Set up the hyperdrive read write interface to access a deployed hyperdrive pool.
-
-    All arguments are fixtures.
-
-    Returns
-    -------
-        HyperdriveReadWriteInterface
-            The interface to access and write to the deployed hyperdrive pool.
-    """
-    rpc_uri = cast(HTTPProvider, _local_hyperdrive_pool.web3.provider).endpoint_uri or URI("http://localhost:8545")
-    hyperdrive_contract_addresses: HyperdriveAddresses = _local_hyperdrive_pool.hyperdrive_contract_addresses
-    eth_config = EthConfig(artifacts_uri="not used", rpc_uri=rpc_uri, abi_dir="./packages/hyperdrive/src/abis")
-    return HyperdriveReadWriteInterface(eth_config, addresses=hyperdrive_contract_addresses)
-
-
-@pytest.fixture(scope="function")
-def hyperdrive_read_write_interface(
-    local_hyperdrive_pool: DeployedHyperdrivePool,
-) -> Iterator[HyperdriveReadWriteInterface]:
-    """Fixture representing a hyperdrive interface to a deployed hyperdrive pool.
-
-    Arguments
-    ---------
-    local_hyperdrive_pool: DeployedHyperdrivePool
-        Fixture representing the deployed hyperdrive pool
-
-    Yields
-    ------
-    HyperdriveReadWriteInterface
-        The interface to access and write to the deployed hyperdrive pool.
-    """
-    yield create_hyperdrive_read_write_interface(local_hyperdrive_pool)
