+++ tests/bot_load_state_test.py
@@ -12,6 +12,7 @@ from chainsync.exec import acquire_data, data_analysis
 from eth_typing import URI
 from ethpy import EthConfig
 from ethpy.hyperdrive.addresses import HyperdriveAddresses
+from ethpy.test_fixtures import DeployedHyperdrivePool
 from fixedpointmath import FixedPoint
 from sqlalchemy.orm import Session
 from web3 import HTTPProvider
@@ -22,7 +23,6 @@ from agent0.base.config import AgentConfig, EnvironmentConfig
 from agent0.hyperdrive import HyperdriveMarketAction, HyperdriveReadInterface, HyperdriveWallet
 from agent0.hyperdrive.exec import setup_and_run_agent_loop
 from agent0.hyperdrive.policies import HyperdriveBasePolicy
-from agent0.hyperdrive.test_fixtures.local_chain import DeployedHyperdrivePool
 
 
 class WalletTestPolicy(HyperdriveBasePolicy):
