+++ tests/bot_wallet_test.py
@@ -11,6 +11,7 @@ from eth_typing import URI
 from ethpy import EthConfig
 from ethpy.hyperdrive import AssetIdPrefix, encode_asset_id
 from ethpy.hyperdrive.addresses import HyperdriveAddresses
+from ethpy.test_fixtures import DeployedHyperdrivePool
 from fixedpointmath import FixedPoint
 from web3 import HTTPProvider
 
@@ -20,7 +21,6 @@ from agent0.base.config import AgentConfig, EnvironmentConfig
 from agent0.hyperdrive import HyperdriveMarketAction, HyperdriveReadInterface, HyperdriveWallet
 from agent0.hyperdrive.exec import setup_and_run_agent_loop
 from agent0.hyperdrive.policies import HyperdriveBasePolicy
-from agent0.hyperdrive.test_fixtures.local_chain import DeployedHyperdrivePool
 
 
 def ensure_agent_wallet_is_correct(wallet: HyperdriveWallet, interface: HyperdriveReadInterface) -> None:
