+++ tests/slippage_warning_test.py
@@ -20,8 +20,7 @@ from agent0.test_utils import CycleTradesPolicy
 
 if TYPE_CHECKING:
     from ethpy.hyperdrive import HyperdriveAddresses
-
-    from agent0.hyperdrive.test_fixtures.local_chain import DeployedHyperdrivePool
+    from ethpy.test_fixtures import DeployedHyperdrivePool
 
 
 class TestSlippageWarning:
